14_Authentications/password_based_authentication/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
